Finished Testing Mode
* The user can now enter testing mode, receive a rocket specification,
then receive a win or loss message depending on whether they met the
specification.

* The test mode displays a win streak that resets to 0 if a user
incorrectly meets a "client's" specification or switches back to the
sandbox

* The user has to attempt to meet a client's specifications before they
can switch back to the sandbox

* ISSUE: The dialog boxes need to be interacted with for the test level
to run smoothly. We want to disable the buttons onscreen while dialogs
are open, so that the user doesn't mess with the simulator before a new
specification is generated. However, if the user hits the X button on
either dialog, it essentially softlocks the application. Ideal solution:
disable the X button on QDialogs, or at least get a signal from them
that can be used to enable the buttons. Possible implementation:
Overriding closeEvent or reject in the dialog classes.